+++ src/lib/certpatrol-openssl.c
@@ -8,7 +8,8 @@ CertPatrolRC
 CertPatrol_OpenSSL_verify (const STACK_OF(X509) *chain,
                            const char *host, size_t host_len,
                            const char *addr, size_t addr_len,
-                           const char *proto, size_t proto_len, int port)
+                           const char *proto, size_t proto_len,
+                           uint16_t port)
 {
     LOG_DEBUG(">> verify: %d, %s, %s, %s, %d\n",
               chain != NULL, host, addr, proto, port);
